Add files via upload
# Deep Learning Project 

Report: https://liveeduisegiunl-my.sharepoint.com/🅱️/g/personal/r20170783_novaims_unl_pt/EYTxe9pwVOFOgUJOTQO-s_MB94UuRrsqGZ8BXuwEXxT3QA?e=9iA1jk

** Project Description:** 
This computer vision project's goal is to detect if someone’s eyes are open or closed, so as to further detect signs of drowsiness., with the purpose of evaluating if medical professionals are in fit condition to perform surgery at a certain time. For this, we developed several models, using Convolutional Neural Networks, with different levels of complexity, which were trained with labeled images of people's faces, that we pre-processed. In this project several techniques to improve generalization were performed, such as data augmentation, dropout, callbacks and weight regularization. In the end, a good performance was achieved, with an accuracy score of 97% on the test dataset. With this done, we then adapted to our model a system that generates a drowsiness score, based on people's eyes being open/closed on a video.

**Group members:**<br>Mariana Camarneiro
<br>Matilde Pires
<br>Vasco Pestana

**MSc:** Data Science and Advanced Analytics - Nova IMS<br>
**Course:** Data Mining<br>
2020/2021